Abstraction of Deduper and an example SimpleEvent Implementation for AppBuilder.
