+++ library/src/main/java/com/datatorrent/lib/bucket/BucketStore.java
@@ -26,7 +26,7 @@ import javax.annotation.Nonnull;
  * @param <T>
  * @since 0.9.4
  */
-public interface BucketStore<T extends Bucketable> extends Cloneable
+public interface BucketStore<T> extends Cloneable
 {
   /**
    * Performs setup operations eg. crate database connections, delete events of windows greater than last committed
@@ -78,9 +78,10 @@ public interface BucketStore<T extends Bucketable> extends Cloneable
    * @param writeEventKeysOnly
    */
   void setWriteEventKeysOnly(boolean writeEventKeysOnly);
+
   BucketStore<T> clone() throws CloneNotSupportedException;
 
-  public interface ExpirableBucketStore<T extends Bucketable & Event> extends BucketStore<T>
+  public interface ExpirableBucketStore<T> extends BucketStore<T>
   {
     void deleteExpiredBuckets(long time) throws IOException;
   }
