+++ library/src/test/java/com/datatorrent/lib/bucket/BucketManagerTest.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2014 DataTorrent, Inc. ALL Rights Reserved.
+ * Copyright (c) 2015 DataTorrent, Inc. ALL Rights Reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -49,7 +49,7 @@ public class BucketManagerTest
   {
 
     @Override
-    public void bucketLoaded(Bucket<DummyEvent> bucket)
+    public void bucketLoaded(AbstractBucket<DummyEvent> bucket)
     {
       try {
         eventBucketExchanger.exchange(bucket.bucketKey);
@@ -84,7 +84,7 @@ public class BucketManagerTest
       manager.newEvent(bucket1, dummyEvent);
     }
     manager.endWindow(round);
-    Bucket<DummyEvent> bucket = manager.getBucket(bucket1);
+    AbstractBucket<DummyEvent> bucket = manager.getBucket(bucket1);
     Assert.assertNotNull(bucket);
     Assert.assertEquals("no of events", (round + 1) * 10, bucket.countOfWrittenEvents());
   }
@@ -126,11 +126,10 @@ public class BucketManagerTest
     manager.loadBucketData(bucket2);
     eventBucketExchanger.exchange(null);
     Assert.assertNotNull(manager.getBucket(bucket2));
-    BucketManagerImpl<DummyEvent> clonedManager = manager.clone();
+    AbstractBucketManager<DummyEvent> clonedManager = manager.clone();
     Assert.assertNull(clonedManager.getBucket(bucket1));
     Assert.assertNotNull(clonedManager.getBucket(bucket2));
     Assert.assertNotNull(clonedManager.getBucketStore());
-    Assert.assertNotNull(clonedManager.bucketStore.equals(manager.bucketStore));
     Assert.assertTrue(clonedManager.writeEventKeysOnly==manager.writeEventKeysOnly);
     Assert.assertTrue(clonedManager.noOfBuckets==manager.noOfBuckets);
     Assert.assertTrue(clonedManager.noOfBucketsInMemory==manager.noOfBucketsInMemory);
